fix: output file per appwindow, #145
